Fixes for ingest completion update.
Accept WAIT_ON_QUEUES as a valid state - a new
possible state while waiting for ingest
completion.

Detect if upload queue deleted and quit.  If the
upload queue is deleted, then the ingest either
completed or was cancelled.  Previously, the
client would think that it needed to renew its
AWS credentials.